Add files via upload
These initial files represent the cut-off (based on federal guidelines) for income eligiblity of gross and net income (excluding BBCE), as well as deductions for standard, shelter, and utility. The utility standards are just for the State of Illinois, for now, because I couldn't find a good federal-level source and this is a data point set by individual states. In 2008, SNAP transitioned to actual monthly costs for dependent child and disabled adult care. Because this is not something typically captured in federal administrative data or surveys, I will use an estimate based on where the value was at in 2008 and inflation.